refactor: consolidate packages under bia_ro_crate namespace
Move bia_ro_crate/, models/, and validator/ into a single bia_ro_crate/
top-level package. Update all absolute imports and the hardcoded module
name filter in get_all_ro_crate_classes() to reflect the new paths.
Update pyproject.toml to declare only the bia_ro_crate package.
