+++ bia-ro-crate/bia_ro_crate/core/parser/ro_crate_metadata_parser.py
@@ -1,0 +1,6 @@
+from bia_ro_crate.core.bia_ro_crate_metadata import BIAROCrateMetadata
+from bia_ro_crate.core.parser.metadata_parser import MetadataParser
+
+
+class ROCrateMetadataParser(MetadataParser[BIAROCrateMetadata]):
+    pass
