+++ bia-ro-crate/bia_ro_crate/models/linked_data/ld_context/ContextTerm.py
