+++ bia-ro-crate/bia_ro_crate/models/linked_data/pydantic_ld/LDModel.py
@@ -1,6 +1,6 @@
 from typing import Annotated
 
-from models.linked_data.ld_context.ContextTerm import ContextTerm
+from bia_ro_crate.models.linked_data.ld_context.ContextTerm import ContextTerm
 from pydantic import AfterValidator, BaseModel, Field
 from rdflib import OWL, RDF
 from rdflib.graph import Graph
