--- bia-ro-crate/bia_ro_crate/parser/ro_crate_metadata_parser.py
@@ -1,6 +1,0 @@
-from bia_ro_crate.bia_ro_crate_metadata import BIAROCrateMetadata
-from bia_ro_crate.parser.metadata_parser import MetadataParser
-
-
-class ROCrateMetadataParser(MetadataParser[BIAROCrateMetadata]):
-    pass
